+++ package.json
@@ -26,11 +26,13 @@
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-standard": "^3.1.0"
+    "eslint-plugin-standard": "^3.1.0",
+    "picgo": "^1.5.0-alpha.0"
   },
   "dependencies": {
     "@1password/connect": "^1.0.5",
-    "@types/notp": "^2.0.2",
-    "thirty-two": "^1.0.2"
+    "@ctrl/ts-base32": "^2.1.1",
+    "@types/node": "^17.0.24",
+    "@types/notp": "^2.0.2"
   }
 }
