--- server.conf
@@ -1,16 +1,0 @@
-server {
-    server_name pic-upload.yuki.xin;
-
-    root /app/PicUploader;
-
-    location / {
-        index index.php;
-        try_files $uri $uri/ index.php$is_args$args;
-    }
-
-    location ~ \.php$ {
-        fastcgi_pass 127.0.0.1:9000;
-        fastcgi_index index.php;
-        include fastcgi.conf;
-    }
-}
